perf(lp): モバイル初期ロードを軽量化しCWVを改善
SP検索順位対策として、LP初回ロードのJS実行量とちらつきを削減する。

- Clerk/ConvexプロバイダをルートからAuthProviders(_auth/_unregistered/認証ページ)へ移動し、公開ページのバンドルから除外
- HeaderのUserMenuを遅延importにし、public variant経由でClerkを巻き込まないよう修正
- デモページを_unregistered外の独立ルートへ移動(Clerk/Convex不要)
- GTM初期化をrequestIdleCallbackまで遅延し、遅延前イベントはバッファしてinit時にflush
- prerender済みDOMの差し替えをflushSyncで同期コミットし、空白フレームのちらつきを解消
- 記事Markdown本文をエントリーチャンクから分離: ?frontmatterプラグイン + articleMeta(メタ専用)を新設し、ルートhead/LPプレビューは本文を読み込まない
- ヒーロー画像にfetchPriority=high/decoding/aspectRatioを付与しLCP前倒し・CLS抑止
- sitemapのルートURLをcanonical(末尾スラッシュ)に統一

エントリーチャンク: 1,236KB→951KB (gzip 338KB→261KB)。Clerk/Convexクライアント/記事本文はLP初期ロードから分離。
