+++ src/components/config/AuthProviders.tsx
@@ -1,0 +1,18 @@
+import { ClerkProvider } from "@clerk/clerk-react";
+import { jaJP } from "@clerk/localizations";
+import type { ReactNode } from "react";
+import { ConvexClientProvider } from "@/src/components/config/ConvexProvider";
+import { CLERK_PUBLISHABLE_KEY, CONVEX_URL } from "@/src/constants/env";
+
+/**
+ * Clerk + Convex をまとめた認証系プロバイダ。
+ * LP・記事・利用規約などの公開ページに Clerk/Convex のバンドルを載せないため、
+ * main.tsx ではなく認証が必要なレイアウト/ページ単位でラップする。
+ * ConvexReactClient はモジュールシングルトン（ConvexProvider.tsx の getClient）なので、
+ * レイアウト間の遷移でプロバイダが再マウントされても WebSocket 接続は維持される。
+ */
+export const AuthProviders = ({ children }: { children: ReactNode }) => (
+  <ClerkProvider publishableKey={CLERK_PUBLISHABLE_KEY} localization={jaJP}>
+    <ConvexClientProvider env={CONVEX_URL}>{children}</ConvexClientProvider>
+  </ClerkProvider>
+);
