+++ src/components/features/ArticleSite/articleContent.ts
@@ -1,12 +1,18 @@
+import { type MarkdownBlock, parseMarkdownBlocks, parseMarkdownDocument } from "@/src/helpers/markdown";
 import {
-  type MarkdownBlock,
-  parseBoundedPositiveInteger,
-  parseMarkdownBlocks,
-  parseMarkdownDocument,
-  parsePositiveInteger,
-  splitList,
-} from "@/src/helpers/markdown";
-import { SITE_URL } from "@/src/helpers/seo";
+  type ArticleMetadata,
+  type CategoryMetadata,
+  createImageSrcResolver,
+  parseArticleMetadata,
+  parseCategoryMetadata,
+  resolveArticleSlug,
+} from "./articleMeta";
+
+/**
+ * 記事サイトの「本文」層。`?raw` で Markdown 本文まで eager import するため、記事詳細・カテゴリ詳細を
+ * 描画するコード分割済みコンポーネント（`ArticleSite/index.tsx`）からのみ import すること。
+ * ルートの `head()` や LP など公開ページの入口からは、本文を含まない `articleMeta.ts` を使う。
+ */
 
 export type {
   MarkdownBlock,
@@ -14,64 +20,24 @@ export type {
   MarkdownImageAlign,
   MarkdownMediaAlign,
 } from "@/src/helpers/markdown";
-
-export type SitePageMetadata = {
-  title: string;
-  description: string;
-  breadcrumbLabel: string;
-  concernTitle: string;
-  latestTitle: string;
-  ctaTitle: string;
-  ctaDescription: string;
-  ctaPrimaryLabel: string;
-  ctaPrimaryHref: string;
-  ctaSecondaryLabel: string;
-  ctaSecondaryHref: string;
-  concernSlugs: string[];
-  landingPreviewTitle: string;
-  landingPreviewDescription: string;
-  landingPreviewLimit: number;
-  landingPreviewLinkLabel: string;
-};
-
-export type CategoryMetadata = {
-  slug: string;
-  title: string;
-  description: string;
-  breadcrumbLabel: string;
-  pointTitle: string;
-  pointDescription: string;
-  concerns: string[];
-  representativeSlug: string;
-  relatedConcernSlugs: string[];
-  ctaTitle: string;
-  ctaDescription: string;
-};
-
-export type ArticleMetadata = {
-  slug: string;
-  title: string;
-  description: string;
-  heroImage?: ArticleHeroImage;
-  publishedAt: string;
-  updatedAt?: string;
-  categorySlug: string;
-  categoryLabel: string;
-  author: string;
-  readingMinutes: number;
-  keywords: string[];
-  relatedSlugs: string[];
-  featured: boolean;
-  canonicalPath: string;
-  ogTitle: string;
-  ogDescription: string;
-};
-
-export type ArticleHeroImage = {
-  src: string;
-  alt: string;
-  width: number;
-};
+export type {
+  ArticleHeroImage,
+  ArticleJsonLd,
+  ArticleMetadata,
+  BreadcrumbJsonLd,
+  CategoryMetadata,
+  ConcernContent,
+  SitePageMetadata,
+} from "./articleMeta";
+export {
+  concerns,
+  createArticleBreadcrumbJsonLd,
+  createArticleJsonLd,
+  createCategoryBreadcrumbJsonLd,
+  getArticleOgpImagePath,
+  parseSitePageMarkdown,
+  sitePage,
+} from "./articleMeta";
 
 export type ArticleContent = {
   meta: ArticleMetadata;
@@ -84,54 +50,6 @@ export type CategoryContent = {
   blocks: MarkdownBlock[];
 };
 
-export type ConcernContent = {
-  slug: string;
-  title: string;
-  description: string;
-  href: string;
-  representativeSlug: string;
-};
-
-export type ArticleJsonLd = {
-  "@context": "https://schema.org";
-  "@type": "BlogPosting";
-  headline: string;
-  description: string;
-  image: string;
-  datePublished: string;
-  dateModified: string;
-  author: {
-    "@type": "Organization";
-    name: string;
-  };
-  publisher: {
-    "@type": "Organization";
-    name: string;
-    logo: {
-      "@type": "ImageObject";
-      url: string;
-    };
-  };
-  mainEntityOfPage: string;
-};
-
-export type BreadcrumbJsonLd = {
-  "@context": "https://schema.org";
-  "@type": "BreadcrumbList";
-  itemListElement: {
-    "@type": "ListItem";
-    position: number;
-    name: string;
-    item?: string;
-  }[];
-};
-
-const pageModules = import.meta.glob<string>("./content/pages/*.md", {
-  eager: true,
-  query: "?raw",
-  import: "default",
-});
-
 const categoryModules = import.meta.glob<string>("./content/categories/*/index.md", {
   eager: true,
   query: "?raw",
@@ -143,58 +61,6 @@ const articleModules = import.meta.glob<string>("./content/articles/*/index.md",
   query: "?raw",
   import: "default",
 });
-
-const imageModules = import.meta.glob<string>("./content/**/*.{avif,gif,jpeg,jpg,png,svg,webp}", {
-  eager: true,
-  query: "?url",
-  import: "default",
-});
-
-const pageRequiredFields = [
-  "title",
-  "description",
-  "breadcrumbLabel",
-  "concernTitle",
-  "latestTitle",
-  "ctaTitle",
-  "ctaDescription",
-  "ctaPrimaryLabel",
-  "ctaPrimaryHref",
-  "ctaSecondaryLabel",
-  "ctaSecondaryHref",
-] as const;
-
-const categoryRequiredFields = [
-  "slug",
-  "title",
-  "description",
-  "breadcrumbLabel",
-  "pointTitle",
-  "pointDescription",
-  "concerns",
-  "representativeSlug",
-  "ctaTitle",
-  "ctaDescription",
-] as const;
-
-const articleRequiredFields = [
-  "title",
-  "description",
-  "publishedAt",
-  "categorySlug",
-  "categoryLabel",
-  "author",
-  "readingMinutes",
-  "canonicalPath",
-  "ogTitle",
-  "ogDescription",
-] as const;
-
-const ARTICLE_HERO_IMAGE_DEFAULT_WIDTH = 320;
-const ARTICLE_HERO_IMAGE_MIN_WIDTH = 240;
-const ARTICLE_HERO_IMAGE_MAX_WIDTH = 360;
-
-export const sitePage = parseSitePageMarkdown(pageModules["./content/pages/articles.md"] ?? "", "articles");
 
 export const categories = Object.entries(categoryModules)
   .map(([path, source]) => {
@@ -210,27 +76,12 @@ export const articles = Object.entries(articleModules)
   })
   .sort((a, b) => b.meta.publishedAt.localeCompare(a.meta.publishedAt));
 
-const articleSlugAliases = {
-  "line-shift-collection-guide": "shiftori-line-workflow",
-} as const;
-
-export const concerns = sitePage.concernSlugs
-  .map((slug) => getCategory(slug))
-  .filter((category): category is CategoryContent => Boolean(category))
-  .map<ConcernContent>((category) => ({
-    slug: category.meta.slug,
-    title: category.meta.title,
-    description: category.meta.description,
-    href: `/articles/categories/${category.meta.slug}`,
-    representativeSlug: category.meta.representativeSlug,
-  }));
-
 export function getArticle(slug?: string): ArticleContent | undefined {
   if (!slug) {
     return articles.find((article) => article.meta.featured);
   }
 
-  const resolvedSlug = articleSlugAliases[slug as keyof typeof articleSlugAliases] ?? slug;
+  const resolvedSlug = resolveArticleSlug(slug);
   return articles.find((article) => article.meta.slug === resolvedSlug);
 }
 
@@ -266,139 +117,19 @@ export function getRepresentativeArticle(category: CategoryContent | undefined):
   return getArticle(category.meta.representativeSlug) ?? getArticlesByCategory(category.meta.slug)[0];
 }
 
-/** 記事別OGP画像のサイトルート相対パス。`pnpm ogp:articles` が同じ場所に生成する。 */
-export function getArticleOgpImagePath(slug: string): string {
-  return `/ogp/articles/${slug}.png`;
-}
-
-export function createArticleJsonLd(article: ArticleContent): ArticleJsonLd {
-  return {
-    "@context": "https://schema.org",
-    "@type": "BlogPosting",
-    headline: article.meta.ogTitle,
-    description: article.meta.ogDescription,
-    image: `${SITE_URL}${getArticleOgpImagePath(article.meta.slug)}`,
-    datePublished: article.meta.publishedAt,
-    dateModified: article.meta.updatedAt ?? article.meta.publishedAt,
-    author: {
-      "@type": "Organization",
-      name: article.meta.author,
-    },
-    publisher: {
-      "@type": "Organization",
-      name: "シフトリ",
-      logo: {
-        "@type": "ImageObject",
-        url: `${SITE_URL}/logo512.png`,
-      },
-    },
-    mainEntityOfPage: `${SITE_URL}${article.meta.canonicalPath}`,
-  };
-}
-
-/** パンくずの先頭（記事一覧トップ）。UIの `Breadcrumbs` と同じ並びを構造化データに写す */
-function articlesRootBreadcrumbItem(): BreadcrumbJsonLd["itemListElement"][number] {
-  return { "@type": "ListItem", position: 1, name: sitePage.breadcrumbLabel, item: `${SITE_URL}/articles` };
-}
-
-export function createArticleBreadcrumbJsonLd(article: ArticleContent): BreadcrumbJsonLd {
-  return {
-    "@context": "https://schema.org",
-    "@type": "BreadcrumbList",
-    itemListElement: [
-      articlesRootBreadcrumbItem(),
-      {
-        "@type": "ListItem",
-        position: 2,
-        name: article.meta.categoryLabel,
-        item: `${SITE_URL}/articles/categories/${article.meta.categorySlug}`,
-      },
-      { "@type": "ListItem", position: 3, name: article.meta.title },
-    ],
-  };
-}
-
-export function createCategoryBreadcrumbJsonLd(category: CategoryContent): BreadcrumbJsonLd {
-  return {
-    "@context": "https://schema.org",
-    "@type": "BreadcrumbList",
-    itemListElement: [
-      articlesRootBreadcrumbItem(),
-      { "@type": "ListItem", position: 2, name: category.meta.breadcrumbLabel },
-    ],
-  };
-}
-
-export function parseSitePageMarkdown(source: string, slug: string): SitePageMetadata {
-  const { frontmatter } = parseMarkdownDocument(source, slug);
-  ensureFields(frontmatter, pageRequiredFields, slug);
-
-  return {
-    title: frontmatter.title,
-    description: frontmatter.description,
-    breadcrumbLabel: frontmatter.breadcrumbLabel,
-    concernTitle: frontmatter.concernTitle,
-    latestTitle: frontmatter.latestTitle,
-    ctaTitle: frontmatter.ctaTitle,
-    ctaDescription: frontmatter.ctaDescription,
-    ctaPrimaryLabel: frontmatter.ctaPrimaryLabel,
-    ctaPrimaryHref: frontmatter.ctaPrimaryHref,
-    ctaSecondaryLabel: frontmatter.ctaSecondaryLabel,
-    ctaSecondaryHref: frontmatter.ctaSecondaryHref,
-    concernSlugs: splitList(frontmatter.concernSlugs),
-    landingPreviewTitle: frontmatter.landingPreviewTitle ?? "シフト作成のヒント",
-    landingPreviewDescription:
-      frontmatter.landingPreviewDescription ??
-      "LINE回収やExcel転記など、シフト作成でつまずきやすいポイントを整理しています。",
-    landingPreviewLimit: parsePositiveInteger(frontmatter.landingPreviewLimit, 3),
-    landingPreviewLinkLabel: frontmatter.landingPreviewLinkLabel ?? "記事一覧を見る",
-  };
-}
-
 export function parseCategoryMarkdown(source: string, slug: string, documentPath?: string): CategoryContent {
-  const { frontmatter, bodySource } = parseMarkdownDocument(source, slug);
-  ensureFields(frontmatter, categoryRequiredFields, slug);
+  const meta = parseCategoryMetadata(source, slug);
+  const { bodySource } = parseMarkdownDocument(source, slug);
 
   return {
-    meta: {
-      slug: frontmatter.slug,
-      title: frontmatter.title,
-      description: frontmatter.description,
-      breadcrumbLabel: frontmatter.breadcrumbLabel,
-      pointTitle: frontmatter.pointTitle,
-      pointDescription: frontmatter.pointDescription,
-      concerns: splitList(frontmatter.concerns),
-      representativeSlug: frontmatter.representativeSlug,
-      relatedConcernSlugs: splitList(frontmatter.relatedConcernSlugs),
-      ctaTitle: frontmatter.ctaTitle,
-      ctaDescription: frontmatter.ctaDescription,
-    },
+    meta,
     blocks: parseMarkdownBlocks(bodySource, { resolveImageSrc: createImageSrcResolver(documentPath) }),
   };
 }
 
 export function parseArticleMarkdown(source: string, slug: string, documentPath?: string): ArticleContent {
-  const { frontmatter, bodySource } = parseMarkdownDocument(source, slug);
-  ensureFields(frontmatter, articleRequiredFields, slug);
-
-  const meta: ArticleMetadata = {
-    slug,
-    title: frontmatter.title,
-    description: frontmatter.description,
-    heroImage: parseArticleHeroImage(frontmatter, documentPath, slug),
-    publishedAt: frontmatter.publishedAt,
-    updatedAt: frontmatter.updatedAt || undefined,
-    categorySlug: frontmatter.categorySlug,
-    categoryLabel: frontmatter.categoryLabel,
-    author: frontmatter.author,
-    readingMinutes: Number(frontmatter.readingMinutes),
-    keywords: splitList(frontmatter.keywords),
-    relatedSlugs: splitList(frontmatter.relatedSlugs),
-    featured: frontmatter.featured === "true",
-    canonicalPath: frontmatter.canonicalPath,
-    ogTitle: frontmatter.ogTitle,
-    ogDescription: frontmatter.ogDescription,
-  };
+  const meta = parseArticleMetadata(source, slug, documentPath);
+  const { bodySource } = parseMarkdownDocument(source, slug);
 
   const blocks = parseMarkdownBlocks(bodySource, { resolveImageSrc: createImageSrcResolver(documentPath) });
   const toc = blocks
@@ -408,76 +139,4 @@ export function parseArticleMarkdown(source: string, slug: string, documentPath?
     .map((block) => ({ id: block.id, text: block.text }));
 
   return { meta, blocks, toc };
-}
-
-function parseArticleHeroImage(
-  frontmatter: Record<string, string>,
-  documentPath: string | undefined,
-  slug: string,
-): ArticleHeroImage | undefined {
-  if (!frontmatter.heroImageSrc) {
-    return undefined;
-  }
-
-  if (!frontmatter.heroImageAlt) {
-    throw new Error(`記事 "${slug}" の heroImageSrc には heroImageAlt が必要です`);
-  }
-
-  return {
-    src: resolveMarkdownImageSrc(frontmatter.heroImageSrc, documentPath),
-    alt: frontmatter.heroImageAlt,
-    width: parseBoundedPositiveInteger(
-      frontmatter.heroImageWidth,
-      ARTICLE_HERO_IMAGE_DEFAULT_WIDTH,
-      ARTICLE_HERO_IMAGE_MIN_WIDTH,
-      ARTICLE_HERO_IMAGE_MAX_WIDTH,
-    ),
-  };
-}
-
-function ensureFields<const T extends readonly string[]>(
-  frontmatter: Record<string, string>,
-  fields: T,
-  slug: string,
-): void {
-  for (const field of fields) {
-    if (!frontmatter[field]) {
-      throw new Error(`記事 "${slug}" の frontmatter に ${field} がありません`);
-    }
-  }
-}
-
-function createImageSrcResolver(documentPath?: string): (src: string) => string {
-  return (src) => resolveMarkdownImageSrc(src, documentPath);
-}
-
-function resolveMarkdownImageSrc(src: string, documentPath?: string): string {
-  if (/^(https?:)?\/\//.test(src) || /^(data|blob):/.test(src) || src.startsWith("/")) {
-    return src;
-  }
-
-  if (!documentPath) {
-    return src;
-  }
-
-  const documentDirectory = documentPath.replace(/\/[^/]*$/, "");
-  const normalizedPath = normalizeContentPath(`${documentDirectory}/${src}`);
-  return imageModules[normalizedPath] ?? src;
-}
-
-function normalizeContentPath(path: string): string {
-  const segments: string[] = [];
-
-  for (const segment of path.split("/")) {
-    if (!segment || segment === ".") {
-      continue;
-    }
-    if (segment === "..") {
-      segments.pop();
-      continue;
-    }
-    segments.push(segment);
-  }
-
-  return `./${segments.join("/")}`;
 }
