+++ src/components/templates/Header/index.tsx
@@ -1,9 +1,14 @@
 import type { BoxProps, ContainerProps, FlexProps, ImageProps, TextProps } from "@chakra-ui/react";
 import { Box, Container, Flex, Image, Link, Text } from "@chakra-ui/react";
 import { Link as RouterLink } from "@tanstack/react-router";
-import type { ReactNode } from "react";
+import { lazy, type ReactNode, Suspense } from "react";
 import { Button } from "@/src/components/ui/Button";
-import { UserMenu, type UserMenuDeleteShopAction } from "./UserMenu";
+import type { UserMenuDeleteShopAction } from "./UserMenu";
+
+// UserMenu は @clerk/clerk-react（SignOutButton）に依存する。静的 import すると
+// variant="public" の LP など Clerk を使わない公開ページのチャンクにも Clerk が
+// 巻き込まれるため、user variant でだけ読み込む遅延 import にしている。
+const UserMenu = lazy(() => import("./UserMenu").then((m) => ({ default: m.UserMenu })));
 
 export const HEADER_HEIGHT = { base: "64px", md: "68px" } as const;
 export const STAFF_CONTENT_MAX_W = "1024px";
@@ -80,7 +85,11 @@ export const Header = (props: HeaderProps = {}) => {
   return (
     <HeaderShell position={props.position ?? "fixed"}>
       <HeaderBrand to="/dashboard" ariaLabel="ダッシュボードへ" showTagline />
-      {props.showUserMenu !== false && <UserMenu tone="light" deleteShopAction={props.deleteShopAction} />}
+      {props.showUserMenu !== false && (
+        <Suspense fallback={null}>
+          <UserMenu tone="light" deleteShopAction={props.deleteShopAction} />
+        </Suspense>
+      )}
     </HeaderShell>
   );
 };
