+++ src/routes/_auth.tsx
@@ -1,5 +1,6 @@
 import { Box } from "@chakra-ui/react";
 import { createFileRoute, Outlet } from "@tanstack/react-router";
+import { AuthProviders } from "@/src/components/config/AuthProviders";
 import { AuthenticatedHeader } from "@/src/components/features/auths/AuthenticatedHeader";
 import { AuthGuard } from "@/src/components/features/auths/AuthGuard";
 import { UnauthenticatedBoundary } from "@/src/components/features/auths/UnauthenticatedBoundary";
@@ -11,11 +12,13 @@ export const Route = createFileRoute("/_auth")({
 
 function RouteComponent() {
   return (
-    <AuthGuard>
-      <UnauthenticatedBoundary>
-        <AuthenticatedLayout />
-      </UnauthenticatedBoundary>
-    </AuthGuard>
+    <AuthProviders>
+      <AuthGuard>
+        <UnauthenticatedBoundary>
+          <AuthenticatedLayout />
+        </UnauthenticatedBoundary>
+      </AuthGuard>
+    </AuthProviders>
   );
 }
