+++ src/routes/_unregistered.tsx
@@ -1,9 +1,14 @@
 import { createFileRoute, Outlet } from "@tanstack/react-router";
+import { AuthProviders } from "@/src/components/config/AuthProviders";
 
 export const Route = createFileRoute("/_unregistered")({
   component: RouteComponent,
 });
 
 function RouteComponent() {
-  return <Outlet />;
+  return (
+    <AuthProviders>
+      <Outlet />
+    </AuthProviders>
+  );
 }
