+++ src/routes/demo.shiftboard.tsx
@@ -2,7 +2,7 @@ import { createFileRoute } from "@tanstack/react-router";
 import { buildLinks, buildMeta } from "@/src/helpers/seo";
 import { DemoShiftBoardRoutePage } from "@/src/pages/demo-shift-board";
 
-export const Route = createFileRoute("/_unregistered/demo/shiftboard")({
+export const Route = createFileRoute("/demo/shiftboard")({
   head: () => ({
     // 採用案: 登録なしで試せる無料デモ｜店長・シフト担当者視点でシフト管理を体験
     //   狙い: 「シフト管理 デモ」「シフト管理 試す」「シフト管理 登録なし」
